// TaskManagementSytem/frontend/src/types/TaskType.ts
import { PriorityEnum } from "../contants/PriorityEnum";
import { TaskStatusEnum } from "../contants/TaskStatusEnum";

export interface TaskBase {
    taskId?: number; // Optional because it might be generated by the server
    taskName: string;
    taskDescription: string;
    userId?: number;
    dueDate: string;
    status: TaskStatusEnum;
    priority: PriorityEnum | string;
    createdAt: string;
  }
  
  export interface TaskWithCategoryId extends TaskBase {
    categoryId: number;
  }
  
  export interface TaskWithCategoryName extends TaskBase {
    categoryName: string;
  }


  export type TaskType = TaskWithCategoryId | TaskWithCategoryName;